fix: dtype handling in backends

+++ ivy/functional/backends/numpy/statistical.py
@@ -65,8 +65,9 @@ def mean(
 ) -> np.ndarray:
     if dtype is not None:
         dtype = ivy.as_native_dtype(dtype)
+        x = x.astype(dtype)
     axis = tuple(axis) if isinstance(axis, list) else axis
-    return np.mean(x, axis=axis, keepdims=keepdims, dtype=dtype, out=out)
+    return np.mean(x, axis=axis, keepdims=keepdims, dtype=x.dtype, out=out)
 
 
 mean.support_native_out = True
